Add tests

+++ src/main/java/flower/store/FlowerPack.java
@@ -3,7 +3,7 @@ package flower.store;
 public class FlowerPack {
     private Flower flower;
     private int quantity;
-    private int price = 0;
+    private double price = 0;
 
     public FlowerPack(Flower flower, int quantity) {
         this.price += flower.getPrice() * quantity;
